pdf_management_system/requirements.txt
# Core dependencies
pandas==2.1.0
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
openpyxl==3.1.2

# Optional dependencies for development
pytest==7.4.2
black==23.7.0